+++ gpsdclient/client.py
@@ -13,26 +13,63 @@ from typing import Any, Dict, Iterable, Union
 # braces. (https://github.com/tfeldmann/gpsdclient/issues/1)
 REGEX_TRAILING_COMMAS = re.compile(r"\s*,\s*}")
 
+FilterType = Union[str, Iterable[str]]
+
+
+def parse_datetime(x: Any) -> Union[Any, datetime]:
+    """
+    tries to convert the input into a `datetime` object if possible.
+    """
+    try:
+        if isinstance(x, float):
+            return datetime.fromtimestamp(x)
+        elif isinstance(x, str):
+            # time zone information can be omitted because gps always sends UTC.
+            return datetime.strptime(x, "%Y-%m-%dT%H:%M:%S.%fZ")
+    except ValueError:
+        pass
+    return x
+
+
+def create_filter_regex(reports: Union[str, Iterable[str]] = set()) -> str:
+    """
+    Dynamically assemble a regular expression to match the given report classes.
+    This way we don't need to parse the json to filter by report.
+    """
+    if isinstance(reports, str):
+        reports = reports.split(",")
+    if reports:
+        classes = set(x.strip().upper() for x in reports)
+        return r'"class":\s?"(%s)"' % "|".join(classes)
+    return r".*"
+
 
 class GPSDClient:
-    def __init__(self, host: str = "127.0.0.1", port: Union[str, int] = "2947", timeout: int = socket.getdefaulttimeout()) -> None:
+    def __init__(
+        self,
+        host: str = "127.0.0.1",
+        port: Union[str, int] = "2947",
+        timeout: Union[float, int, None] = None,
+    ):
         self.host = host
         self.port = port
         self.timeout = timeout
         self.sock = None  # type: Any
 
-    def json_stream(self, filter: Iterable[str] = set()) -> Iterable[str]:
-        # Dynamically assemble a regular expression to match the given report classes.
-        # This way we don't need to parse the json to filter by report.
-        if filter:
-            report_classes = set(f.strip().upper() for f in filter)
-            filter_regex = re.compile(r'"class":\s?"(%s)"' % "|".join(report_classes))
-
+    def gpsd_lines(self):
         self.close()
-        self.sock = socket.create_connection(address=(self.host, int(self.port)), timeout=self.timeout)
+        self.sock = socket.create_connection(
+            address=(self.host, int(self.port)),
+            timeout=self.timeout,
+        )
         self.sock.send(b'?WATCH={"enable":true,"json":true}\n')
+        yield from self.sock.makefile("r", encoding="utf-8")
+
+    def json_stream(self, filter: FilterType = set()) -> Iterable[str]:
+        filter_regex = re.compile(create_filter_regex(filter))
+
         expect_version_header = True
-        for line in self.sock.makefile("r", encoding="utf-8"):
+        for line in self.gpsd_lines():
             answ = line.strip()
             if answ:
                 if expect_version_header and not answ.startswith('{"class":"VERSION"'):
@@ -48,26 +85,13 @@ class GPSDClient:
                     yield cleaned_json
 
     def dict_stream(
-        self, *, convert_datetime: bool = True, filter: Iterable[str] = set()
+        self, *, convert_datetime: bool = True, filter: FilterType = set()
     ) -> Iterable[Dict[str, Any]]:
         for line in self.json_stream(filter=filter):
             result = json.loads(line)
             if convert_datetime and "time" in result:
-                result["time"] = self._convert_datetime(result["time"])
+                result["time"] = parse_datetime(result["time"])
             yield result
-
-    @staticmethod
-    def _convert_datetime(x: Any) -> Union[Any, datetime]:
-        """converts the input into a `datetime` object if possible."""
-        try:
-            if isinstance(x, float):
-                return datetime.fromtimestamp(x)
-            elif isinstance(x, str):
-                # time zone information can be omitted because gps always sends UTC.
-                return datetime.strptime(x, "%Y-%m-%dT%H:%M:%S.%fZ")
-        except ValueError:
-            pass
-        return x
 
     def close(self):
         if self.sock:
@@ -76,6 +100,12 @@ class GPSDClient:
 
     def __str__(self):
         return "<GPSDClient(host=%s, port=%s)>" % (self.host, self.port)
+
+    def __enter__(self):
+        return self
+
+    def __exit__(self, exc_type, exc_value, exc_traceback):
+        self.close()
 
     def __del__(self):
         self.close()